Fixes. we need two versions of binary_op
