+++ flox/aggregations.py
@@ -8,6 +8,7 @@ from functools import cached_property, partial
 from typing import TYPE_CHECKING, Any, Callable, Literal, TypedDict
 
 import numpy as np
+import pandas as pd
 from numpy.typing import ArrayLike, DTypeLike
 
 from . import aggregate_flox, aggregate_npg, xrutils
@@ -63,6 +64,9 @@ def generic_aggregate(
     dtype=None,
     **kwargs,
 ):
+    if func == "identity":
+        return array
+
     if engine == "flox":
         try:
             method = getattr(aggregate_flox, func)
@@ -574,7 +578,7 @@ class Scan:
     # between reductions and scans
     name: str
     # binary operation (e.g. add)
-    # binary_op: Callable
+    binary_op: Callable
     # in-memory grouped scan function (e.g. cumsum)
     scan: str
     # Grouped reduction that yields the last result of the scan (e.g. sum)
@@ -609,9 +613,62 @@ class ScanState:
         assert (self.state is not None) or (self.result is not None)
 
 
-cumsum = Scan("cumsum", reduction="sum", scan="cumsum", identity=0)
-nancumsum = Scan("nancumsum", reduction="nansum", scan="nancumsum", identity=0)
-ffill = Scan("ffill", reduction="nanlast", scan="ffill", identity=np.nan)
+def binary_op(left: AlignedArrays, right: AlignedArrays, *, agg: Scan) -> AlignedArrays:
+    """Implements groupby binary operation. Used for scan."""
+    from .core import reindex_
+
+    reindexed = reindex_(
+        left.array,
+        from_=pd.Index(left.group_idx),
+        # can't use right.group_idx since we need to do the indexing later
+        to=pd.RangeIndex(right.group_idx.max() + 1),
+        fill_value=agg.identity,
+        axis=-1,
+    )
+    return AlignedArrays(
+        array=agg.binary_op(reindexed[..., right.group_idx], right.array), group_idx=right.group_idx
+    )
+
+
+def scan_binary_op(left_state: ScanState, right_state: ScanState, *, agg: Scan) -> ScanState:
+    """
+    Implements the binary op portion of the scan as a concatenate-then-scan.
+    This is useful for `ffill`, and presumably more generalized scans.
+    """
+    from flox.core import grouped_reduce
+
+    assert left_state.state is not None
+    left = left_state.state
+    right = right_state.result if right_state.result is not None else right_state.state
+
+    new_group_idx = np.concatenate([left.group_idx, right.group_idx], axis=-1)
+    new_array = np.concatenate([left.array, right.array], axis=-1)
+
+    new = generic_aggregate(
+        new_group_idx,
+        new_array,
+        func=agg.scan,
+        axis=right.array.ndim - 1,
+        engine="flox",
+        fill_value=agg.identity,
+    )[..., left.group_idx.size :]
+
+    # This is quite important. We need to update the state seen so far and propagate that.
+    lasts = grouped_reduce(
+        AlignedArrays(group_idx=new_group_idx, array=new_array), agg=agg, axis=right.array.ndim - 1
+    )
+    return ScanState(
+        state=lasts.state,
+        result=AlignedArrays(array=new, group_idx=right.group_idx),
+    )
+
+
+cumsum = Scan("cumsum", binary_op=binary_op, reduction="sum", scan="cumsum", identity=0)
+nancumsum = Scan("nancumsum", binary_op=binary_op, reduction="nansum", scan="nancumsum", identity=0)
+# ffill uses the identity for scan, and then at the binary-op state,
+# we concatenate the blockwise-reduced values with the original block,
+# and then execute the scan
+ffill = Scan("ffill", binary_op=scan_binary_op, reduction="nanlast", scan="ffill", identity=np.nan)
 # cumprod = Scan("cumprod", binary_op=np.multiply, preop="prod", scan="cumprod")
