+++ flox/core.py
@@ -2718,7 +2718,7 @@ def groupby_scan(
 
     if not has_dask:
         (single_axis,) = axis_
-        final_state = grouped_scan(
+        final_state = chunk_scan(
             AlignedArrays(array=array, group_idx=by_), axis=single_axis, agg=agg, dtype=agg.dtype
         )
         return extract_array(final_state)
@@ -2726,9 +2726,7 @@ def groupby_scan(
         return dask_groupby_scan(array, by_, axes=axis_, agg=agg)
 
 
-def grouped_scan(
-    inp: AlignedArrays, *, axis: int, agg: Scan, dtype=None, keepdims=None
-) -> ScanState:
+def chunk_scan(inp: AlignedArrays, *, axis: int, agg: Scan, dtype=None, keepdims=None) -> ScanState:
     assert axis == inp.array.ndim - 1
 
     # TODO: factorize here (maybe?)
@@ -2762,34 +2760,7 @@ def grouped_reduce(inp: AlignedArrays, *, agg: Scan, axis: int, keepdims=None) -
     )
 
 
-def scan_binary_op(left_state: ScanState, right_state: ScanState, *, agg: Scan) -> ScanState:
-
-    assert left_state.state is not None
-    left = left_state.state
-    right = right_state.result if right_state.result is not None else right_state.state
-
-    new_group_idx = np.concatenate([left.group_idx, right.group_idx], axis=-1)
-    new_array = np.concatenate([left.array, right.array], axis=-1)
-
-    new = generic_aggregate(
-        new_group_idx,
-        new_array,
-        func=agg.scan,
-        axis=right.array.ndim - 1,
-        engine="flox",
-        fill_value=agg.identity,
-    )[..., left.group_idx.size :]
-    # This is quite important. We need to update the state seen so far and propagate that.
-    lasts = grouped_reduce(
-        AlignedArrays(group_idx=new_group_idx, array=new_array), agg=agg, axis=right.array.ndim - 1
-    )
-    return ScanState(
-        state=lasts.state,
-        result=AlignedArrays(array=new, group_idx=right.group_idx),
-    )
-
-
-def _zip(group_idx, array) -> AlignedArrays:
+def _zip(group_idx: np.ndarray, array: np.ndarray) -> AlignedArrays:
     return AlignedArrays(group_idx=group_idx, array=array)
 
 
@@ -2812,11 +2783,11 @@ def dask_groupby_scan(array, by, axes: T_Axes, agg: Scan) -> DaskArray:
         _zip, by, array, dtype=array.dtype, meta=array._meta, name="groupby-scan-preprocess"
     )
 
-    scan_ = partial(grouped_scan, agg=agg)
+    scan_ = partial(chunk_scan, agg=agg)
     # dask tokenizing error workaround
     scan_.__name__ = scan_.func.__name__
 
-    binop = partial(scan_binary_op, agg=agg)
+    binop = partial(agg.binary_op, agg=agg)
 
     # 2. Run the scan
     accumulated = scan(
